fix: centralize asset url credential stripping
